+++ src/titiler/application/tests/routes/test_cog.py
@@ -479,7 +479,7 @@ def test_validate_cog(app, url):
 
 
 @patch("rio_tiler.io.cogeo.rasterio")
-def test_info_nan(rio, app):
+def test_json_response_with_nan(rio, app):
     """test /info endpoint."""
     rio.open = mock_rasterio_open
 
@@ -488,6 +488,7 @@ def test_info_nan(rio, app):
     body = response.json()
     assert body["dtype"] == "float32"
     assert body["nodata_type"] == "Nodata"
+    assert body["nodata_value"] is None
 
     response = app.get("/cog/info.geojson?url=https://myurl.com/cog_with_nan.tif")
     assert response.status_code == 200
@@ -496,3 +497,16 @@ def test_info_nan(rio, app):
     assert body["geometry"]
     assert body["properties"]["nodata_type"] == "Nodata"
     assert body["properties"]["nodata_value"] is None
+
+    response = app.get("/cog/metadata?url=https://myurl.com/cog_with_nan.tif")
+    assert response.status_code == 200
+    body = response.json()
+    assert body["nodata_type"] == "Nodata"
+    assert body["nodata_value"] is None
+
+    response = app.get(
+        "/cog/point/79.80860440702253,21.852217086223234?url=https://myurl.com/cog_with_nan.tif"
+    )
+    assert response.status_code == 200
+    body = response.json()
+    assert body["values"][0] is None