+++ CHANGES.md
@@ -4,12 +4,12 @@
 
 ### titiler.core
 
-- add custom JSONResponse using [simplejson](https://simplejson.readthedocs.io/en/latest/) to allow NaN/inf/-inf values
+- add custom JSONResponse using [simplejson](https://simplejson.readthedocs.io/en/latest/) to allow NaN/inf/-inf values (ref: https://github.com/developmentseed/titiler/pull/374)
+- use `titiler.core.resources.responses.JSONResponse` as default response for `info`, `metadata`, `statistics` and `point` endpoints (ref: https://github.com/developmentseed/titiler/pull/374)
 
 ### titiler.application
 
 - switch to `starlette_cramjam` compression middleware (ref: https://github.com/developmentseed/titiler/issues/369)
-- use `titiler.core.resources.responses.JSONResponse` as default application response.
 
 ## 0.3.9 (2021-09-07)